Refactored code by seperating routers and logic into routes and controllers directories
